requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aubio==0.4.9
    # via -r requirements.in
bidict==0.22.1
    # via python-socketio
blinker==1.7.0
    # via flask
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-socketio
flask-socketio==5.3.6
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via wsproto
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
numpy==1.26.3
    # via
    #   -r requirements.in
    #   aubio
    #   scipy
packaging==23.2
    # via gunicorn
pyaudio==0.2.14
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-engineio==4.8.0
    # via python-socketio
python-socketio==5.10.0
    # via flask-socketio
scipy==1.11.4
    # via -r requirements.in
simple-websocket==1.0.0
    # via python-engineio
werkzeug==3.0.1
    # via flask
wsproto==1.2.0
    # via simple-websocket
